Correct test fillon by changing date
The test was written for year 2011, which required changing the SMIC's value to 9
through a reform, since the SMIC changes during the year 2011, and the test is simpler with a unique value.
To make the test pass, the year is changed to 2013 in which the SMIC did
not change.
This DOES NOT resolve the initial bug in reforms.update_legislation()
that is the reason the test was failing.